Atualizado pedido de venda

+++ src/main/java/br/com/geremiasemerim/produtos/model/dto/ItemQuantidade.java
@@ -1,6 +1,6 @@
 package br.com.geremiasemerim.produtos.model.dto;
 
 public record ItemQuantidade(
-        int item,
+        Long item,
         int quantidade) {
 }
